requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
dateparser==1.2.1
frozenlist==1.6.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
packaging==25.0
pluggy==1.6.0
propcache==0.3.1
pycryptodome==3.23.0
pytest==8.3.5
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0
